Add tests for eventOptionsKey
I've been impressed with the recent advances that Jest has made, so I
decided to go with that as my testing framework here. It is nice to have
something that just works so easily right out of the box.

To get things started, I decided to test the eventOptionsKey hashing
function I wrote. I admit that I probably over-tested this a bit, but it
will be nice to have these tests I think. I first approached these tests
by asserting the values returned by the function, but I decided that
tested the implementation details too much so I went with a
self-comparative approach.

In doing this testing, I discovered a small problem with the
implementation: we want to treat empty objects the same as false, but
weren't. I fixed that in this commit.