Add tests for addEventListener and removeEventListener
This is the public API, so we want to make sure that it is well-tested.
In the course of writing these tests, I uncovered some issues that I
fixed as part of this PR.

The first is that I had messed up the order of the arguments passed to
TargetEventListeners#add().

The second is I found a small edge case where the underlying
removeEventListener on the target would be called if our
removeEventListener is called for an event that was not previously
added. This is probably not a big deal, but it seems cleaner to avoid
this and any other side-effects that might occur down the road.