Install babel-jest and reinstate Node 4 builds
I was seeing a cryptic error under Node 4:

```
Test suite failed to run

SyntaxError: Unexpected reserved word

  at transformAndBuildScript
  (node_modules/jest/node_modules/jest-cli/node_modules/jest-runtime/build/transform.js:316:10)
  at process._tickCallback (node.js:385:9)
```

I traced it back to this line in Jest:

```
return new vm.Script(wrappedResult, {displayErrors: true, filename});
```

In trying to come up with a minimal reproduction of this issue, I
determined that it happened when using `import` in any file, but not
`require`. It turned out that I needed to bring in babel-jest to make my
tests work in older node versions. Sadly, this error message and line
number were not very helpful in tracking this down. Perhaps this error
message could be made better, or a comment could be added near this line
in jest-runtime that would help point people in the right direction.