Pass event options to handleEvent()
I made an error when extracting this code from react-waypoint that
caused it to not actually work. Since we now rely on the event options
to get the event handlers, we need to have them available
everywhere--including the handleEvent method. I've added some regression
tests to cover this behavior, even though this is technically private
functionality.