Convert TargetEventHandlers from class to function
The code added by Babel to compile this down to ES5 adds about 1.1 KiB
to the built version. We can pretty trivially avoid this for now by
using a function instead of a class. This drops the built filesize from
7.3 KiB to 6.2 KiB, which will be a small win for performance.

I think Babel 7 aims to make some changes that will help us use the
class syntax while avoiding this unnecessary overhead, at which point we
might want to re-evaluate this change. But for now, let's take the easy
bundle size win.