Remove deprecated removeEventListener export
Since the next release of consolidated-events will be semver major, it
is a good time to remove this deprecated named export.

At this time there is now only one named export in this file, which the
linter suggests I should convert to a default export. However, I think
I'd like to minimize the work needed to be done by folks who are
upgrading, so I'm going to leave this as a named export for now.